third_party/blink/web_tests/accessibility/aria-used-on-image-maps-expected.txt

This tests that you can set an ARIA role on image map elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS group.childAtIndex(0).role is 'AXRole: AXButton'
PASS group.childAtIndex(1).role is 'AXRole: AXButton'
PASS successfullyParsed is true

TEST COMPLETE
